+++ lib/src/importer/filesystem.dart
@@ -18,7 +18,7 @@ class FilesystemImporter extends Importer {
   final String _loadPath;
 
   /// Creates an importer that loads files relative to [loadPath].
-  FilesystemImporter(this._loadPath);
+  FilesystemImporter(String loadPath) : _loadPath = p.absolute(loadPath);
 
   Uri canonicalize(Uri url) {
     if (url.scheme != 'file' && url.scheme != '') return null;