+++ tool/grind.dart
@@ -5,6 +5,7 @@
 import 'package:grinder/grinder.dart';
 
 import 'grind/npm.dart';
+import 'grind/standalone.dart';
 import 'grind/synchronize.dart';
 
 export 'grind/bazel.dart';
@@ -35,5 +36,5 @@ format() {
 npmInstall() => run("npm", arguments: ["install"]);
 
 @Task('Runs the tasks that are required for running tests.')
-@Depends(format, synchronize, npmPackage, npmInstall)
+@Depends(format, synchronize, npmPackage, npmInstall, appSnapshot)
 beforeTest() {}